Add an interval class and refactor package
